requirements copy.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
Flask-WTF
email_validator
pymysql
python-dotenv